Unifies upload utilities and streamlines dataset submission
Refactors upload logic by consolidating payload building, TUS file uploads, and API submission into a single, reusable module. Eliminates legacy code and replaces scattered upload functions with a unified API, improving maintainability and testability.
