+++ client/src/components/Visualizations/VisualizationExamples.test.js
@@ -6,13 +6,11 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { ref } from "vue";
 
 import { useToast } from "@/composables/toast";
-import { buildUploadPayload, createUrlUploadItem } from "@/utils/uploadPayload";
-import { sendPayload } from "@/utils/uploadSubmit";
+import { createUrlUploadItem, uploadDatasets } from "@/utils/upload";
 
 import UploadExamples from "./VisualizationExamples.vue";
 
-vi.mock("@/utils/uploadPayload", () => ({
-    buildUploadPayload: vi.fn(() => "mockedPayload"),
+vi.mock("@/utils/upload", () => ({
     createUrlUploadItem: vi.fn((url, historyId, options) => ({
         src: "url",
         url,
@@ -20,10 +18,7 @@ vi.mock("@/utils/uploadPayload", () => ({
         name: options?.name ?? "default",
         ext: options?.ext ?? "auto",
     })),
-}));
-
-vi.mock("@/utils/uploadSubmit", () => ({
-    sendPayload: vi.fn(),
+    uploadDatasets: vi.fn(),
 }));
 
 vi.mock("@/composables/toast");
@@ -89,20 +84,22 @@ describe("UploadExamples.vue", () => {
             name: "Example 1",
             ext: urlData[0].ftype,
         });
-        expect(buildUploadPayload).toHaveBeenCalledWith([
-            expect.objectContaining({
-                src: "url",
-                url: urlData[0].url,
-                historyId: "fake-history-id",
-                name: "Example 1",
-                ext: urlData[0].ftype,
-            }),
-        ]);
-        expect(sendPayload).toHaveBeenCalledWith("mockedPayload", {
-            success: expect.any(Function),
-            error: expect.any(Function),
-        });
-        vi.mocked(sendPayload).mock.calls[0][1].success();
+        expect(uploadDatasets).toHaveBeenCalledWith(
+            [
+                expect.objectContaining({
+                    src: "url",
+                    url: urlData[0].url,
+                    historyId: "fake-history-id",
+                    name: "Example 1",
+                    ext: urlData[0].ftype,
+                }),
+            ],
+            {
+                success: expect.any(Function),
+                error: expect.any(Function),
+            },
+        );
+        vi.mocked(uploadDatasets).mock.calls[0][1].success();
         expect(toastSuccess).toHaveBeenCalledWith("The sample dataset 'Example 1' is being uploaded to your history.");
     });
 
@@ -113,7 +110,7 @@ describe("UploadExamples.vue", () => {
         });
         const items = wrapper.findAllComponents(BDropdownItem);
         await items.at(1).find("a").trigger("click");
-        vi.mocked(sendPayload).mock.calls[0][1].error();
+        vi.mocked(uploadDatasets).mock.calls[0][1].error();
         expect(toastError).toHaveBeenCalledWith("Uploading the sample dataset 'Example 2' has failed.");
     });
